tests/pyproject.toml
[project]
name = "tests"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "Leon" }]
dependencies = ["pytest-cov>=7.0.0", "pytest-docker>=3.2.5", "mimesis>=19.1.0", "httpx>=0.28.1", "pytest-loguru>=0.4.0"]
requires-python = ">=3.12"
readme = "README.md"


[project.optional-dependencies]
dev = [
    "ipykernel>=7.0.1",
    "ruff>=0.14.14",
    "ty>=0.0.14",
]
[tool.pdm]
distribution = false


[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_level = "INFO"
addopts = """
--exitfirst
--pdb
--pdbcls=IPython.terminal.debugger:TerminalPdb
-v
-r a
--cov
--cov-report=term
--cov-report=html
--cov-report=xml
--no-cov-on-fail
"""


[tool.ty.terminal]
output-format = "concise"


[tool.ruff]
output-format = "grouped"


[tool.ruff.lint]
ignore = [
    "E701", # Multiple statements on one line (colon)
    # "F541", # Mf-string without any placeholders
    "F401", # XXX imported but unused
    # "E712", # Avoid equality comparisons to `True`
]